requirements.txt
deepeval
pytest
openai
pinecone-client
python-dotenv
streamlit
PyMuPDF